refactor : 좌표 자료형을 Double로 변경

+++ domain/src/main/java/com/gta/domain/model/Coordinate.kt
@@ -1,6 +1,6 @@
 package com.gta.domain.model
 
 data class Coordinate(
-    val x: Long = 0,
-    val y: Long = 0
+    val x: Double = 0.0,
+    val y: Double = 0.0
 )
